+++ backend/src/pragma/__init__.py
@@ -10,4 +10,4 @@ Returns:
 Raises:
     None.
 """
-__version__ = "0.8.0"
+__version__ = "0.8.1"